Added HTML demo of Flask server
I chained together some of our API requests in Flask for our deliverable for Friday, and it should satisfy the requirements of
(1) a form (2) zip code entry (3) geocoding API (4) NWS API (5) Spoonacular API
I still need to work on (6) Display recipe, but I think we may need to just have it as a hyperlink.
Either way, I think we should work on cleaning up our repository on Friday and getting on the same page for our framework/architecture, as well as our plans for the front end/database. Cheers!